try to use bluetooth-serial-port
