+++ test/test.js
@@ -24,7 +24,7 @@ if (mode === "bt" || mode === "bluetooth") {
         }
     }, 3000);*/
 } else {
-    device.startConnection("com", ["COM4", "COM5", "COM6"]);
+    device.startConnection("com", ["COM3", "COM4", "COM5"]);
 
     device.on("connect", (trackerName) => {
         console.log(`Connected to tracker ${trackerName}`);
